+++ mocks/mockEnv.ts
@@ -4,3 +4,7 @@ export const FULFILL_PATH = "/fulfill/12345";
 export const LIMITED_ACCESS_WORK_PATH =
   "/work/12345678-1234-1234-1234-1234567890ab";
 export const DOWNLOAD_PATH = "/test-download-pdf";
+export const HOME_PATH = "/";
+export const COLLECTION_LIST_PATH = "/collection/list";
+export const COLLECTION_PATH =
+  "/collection/978ea0e0-8ecc-4de2-bfe8-032fea641d8e";